+++ src/Entity/Interfaces/EntityInterface.php
@@ -9,4 +9,12 @@ interface EntityInterface
      * May only return `null` for unstored entities.
      */
     public function getId () : ?int;
+
+
+    /**
+     * Returns whether this entity was already persisted and flushed (`false`) or if it is new (`true`).
+     *
+     * @return bool true if not yet flushed, false otherwise
+     */
+    public function isNew () : bool;
 }